Add empty interface for model classes for generic typing
This new blank interface is only being used for generic typing purposes.
It will allow for methods to be created whose parameters can accept any
one of the model package objects. The interface name probably could use
some creative help. Hopefully I'll think of a name for it that better
describes it's implementers.