Add 2 new read methods to all Mapper classes
Two methods have been added to the Mapper classes in order to give
objects, calling upon a Mapper class, the option of fetching a single
DatabaseItem from the database - as opposed to calling readAll() and
iterating over the resulting list. One of the new methods accepts a
String argument, representing the unique name of the DatabaseItem to be
fetched, and returns the DatabaseItem, found within the database, whose
unique name matches the String given. The other new method accepts an
int argument, representing the ID of the DatabaseItem to be fetched,
and returns the DatabaseItem, found within the database, whose ID
matches the int given.