Update MainViewController.submit()
The method now writes the entire location to the log file as well as the
first name, last name, and display name of the supervisor. The submit
method will no longer include blank/null fields in the log file. The
date and hours field should never be blank and are ALWAYS included.