Autopopulate AddSupervisor display name with first name + last name
When the user is adding a new supervisor, the display name is
autofilled with the first name and the last name of the supervisor as
the user types in the first/last name textfields. The user is still
free to edit the display name after they're finished typing in the
first name and last name.