Cause a log to be submitted or a new tab to be created using enter key
If the user presses the ENTER key within the Hours Worked textfield, the
submit method will be called. If the user presses the enter key when the
Create New Log button has focus - which it does by default now, then it
will activate the button just the same as if it were clicked.