Remove code related to Hours Worked log file
The log file is no longer the primary method for displaying log entries
to the user. This is now done through the pdf export feature which
exports a pdf containing all log entries in a tabular format. The log
file also no longer serves as the database for log entries. Instead,
the SQLite database, which stores companies, locations and supervisors,
now stores log entries created by the user. Therefore, there is no
need to keep any code nor dependencies related to the generation of a
log file nor the generation of a pdf of a log file.